# Conflicts:
#	stp-domain/src/main/java/com/sytoss/domain/bom/analytics/AnalyticDrop.java
#	stp-domain/src/main/java/com/sytoss/domain/bom/analytics/AnalyticGrade.java
#	stp-domain/src/main/java/com/sytoss/domain/bom/analytics/SummaryGrade.java
#	stp-ms-lessons/src/main/java/com/sytoss/lessons/connectors/AnalyticsConnector.java
#	stp-ms-lessons/src/main/java/com/sytoss/lessons/controllers/AnalyticsController.java
#	stp-ms-lessons/src/main/java/com/sytoss/lessons/convertors/AnalyticsConvertor.java
#	stp-ms-lessons/src/main/java/com/sytoss/lessons/services/AnalyticsService.java
#	stp-ms-lessons/src/test/java/com/sytoss/lessons/bdd/given/AnalyticsGiven.java
#	stp-ms-lessons/src/test/java/com/sytoss/lessons/bdd/then/AnalyticsThen.java
#	stp-ms-lessons/src/test/java/com/sytoss/lessons/bdd/when/AnalyticsWhen.java
#	stp-ms-lessons/src/test/resources/features/analytics.feature
#	stp-ms-test-producer/src/main/java/com/sytoss/producer/services/PersonalExamService.java

+++ stp-domain/src/main/java/com/sytoss/domain/bom/analytics/Rating.java
@@ -7,5 +7,5 @@ import lombok.Setter;
 @Setter
 public class Rating extends AnalyticDrop {
 
-    private int rank;
+    private Long rank;
 }
